Fix Add button placement
