+++ app/src/main/java/com/example/mygymapp/ui/pages/ExerciseManagementScreen.kt
@@ -10,12 +10,7 @@ import androidx.compose.runtime.*
 import androidx.compose.runtime.livedata.observeAsState
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.text.AnnotatedString
-import androidx.compose.ui.text.SpanStyle
-import androidx.compose.ui.text.buildAnnotatedString
 import androidx.compose.ui.text.font.FontFamily
-import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
 import androidx.lifecycle.viewmodel.compose.viewModel
 import androidx.navigation.NavController
@@ -23,13 +18,12 @@ import com.example.mygymapp.data.Exercise
 import com.example.mygymapp.model.ExerciseCategory
 import com.example.mygymapp.model.MuscleGroup
 import com.example.mygymapp.ui.components.AddEditExerciseSheet
-import androidx.compose.ui.text.withStyle
 import androidx.compose.foundation.ExperimentalFoundationApi
 import com.example.mygymapp.viewmodel.ExerciseViewModel
-import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.foundation.layout.FlowRow
 import androidx.compose.foundation.layout.ExperimentalLayoutApi
+import com.example.mygymapp.ui.components.ExerciseCardWithHighlight
 
 @OptIn(ExperimentalFoundationApi::class, ExperimentalMaterial3Api::class, ExperimentalLayoutApi::class)
 @Composable
@@ -59,23 +53,21 @@ fun ExerciseManagementScreen(navController: NavController) {
     val collapsedStates = remember { mutableStateMapOf<String, Boolean>() }
 
     Surface(
-        modifier = Modifier.fillMaxSize(),
+        modifier = Modifier
+            .fillMaxSize()
+            .navigationBarsPadding(),
         color = MaterialTheme.colorScheme.background
     ) {
-        Column(modifier = Modifier.fillMaxSize().padding(16.dp)) {
-            Row(
-                modifier = Modifier.fillMaxWidth(),
-                horizontalArrangement = Arrangement.SpaceBetween,
-                verticalAlignment = Alignment.CenterVertically
+        Box(modifier = Modifier.fillMaxSize()) {
+            Column(
+                modifier = Modifier
+                    .fillMaxSize()
+                    .padding(16.dp)
             ) {
                 Text(
                     text = "Exercise Library",
                     style = MaterialTheme.typography.headlineSmall.copy(fontFamily = FontFamily.Serif)
                 )
-                Button(onClick = { editing = null; showSheet = true }) {
-                    Text("➕ Add", fontFamily = FontFamily.Serif)
-                }
-            }
 
             Spacer(modifier = Modifier.height(8.dp))
 
@@ -110,62 +102,81 @@ fun ExerciseManagementScreen(navController: NavController) {
 
             Spacer(modifier = Modifier.height(8.dp))
 
-            if (searchQuery.isNotBlank() || selectedMuscleGroup != null) {
-                if (filteredExercises.isEmpty()) {
-                    Text(
-                        text = "No exercises found.",
-                        modifier = Modifier.padding(16.dp),
-                        style = MaterialTheme.typography.bodyMedium
-                    )
-                } else {
-                    LazyColumn(verticalArrangement = Arrangement.spacedBy(12.dp)) {
-                        items(filteredExercises) { ex ->
-                            ExerciseCardWithHighlight(ex, normalizedQuery, onEdit = {
-                                editing = ex
-                                showSheet = true
-                            }, onDelete = {
-                                vm.delete(ex.id)
-                            })
-                        }
-                    }
-                }
-            } else {
-                LazyColumn(
-                    modifier = Modifier.fillMaxSize(),
-                    verticalArrangement = Arrangement.spacedBy(12.dp)
-                ) {
-                    groupedExercises.forEach { (muscleGroup, items) ->
-                        val isCollapsed = collapsedStates[muscleGroup] ?: true
-                        stickyHeader {
-                            Surface(
-                                modifier = Modifier
-                                    .fillMaxWidth()
-                                    .clickable {
-                                        collapsedStates[muscleGroup] = !isCollapsed
+            Box(modifier = Modifier.weight(1f)) {
+                if (searchQuery.isNotBlank() || selectedMuscleGroup != null) {
+                    if (filteredExercises.isEmpty()) {
+                        Text(
+                            text = "No exercises found.",
+                            modifier = Modifier.padding(16.dp),
+                            style = MaterialTheme.typography.bodyMedium
+                        )
+                    } else {
+                        LazyColumn(
+                            modifier = Modifier.fillMaxSize(),
+                            verticalArrangement = Arrangement.spacedBy(12.dp)
+                        ) {
+                            items(filteredExercises) { ex ->
+                                ExerciseCardWithHighlight(
+                                    ex,
+                                    normalizedQuery,
+                                    onEdit = {
+                                        editing = ex
+                                        showSheet = true
                                     },
-                                color = MaterialTheme.colorScheme.surface,
-                                tonalElevation = 4.dp
-                            ) {
-                                Text(
-                                    text = if (isCollapsed) "▸ $muscleGroup" else "▾ $muscleGroup",
-                                    modifier = Modifier.padding(vertical = 8.dp, horizontal = 16.dp),
-                                    style = MaterialTheme.typography.titleMedium.copy(fontFamily = FontFamily.Serif)
+                                    onDelete = { vm.delete(ex.id) }
                                 )
                             }
                         }
-                        if (!isCollapsed) {
-                            items(items) { ex ->
-                                ExerciseCardWithHighlight(ex, "", onEdit = {
-                                    editing = ex
-                                    showSheet = true
-                                }, onDelete = {
-                                    vm.delete(ex.id)
-                                })
+                    }
+                } else {
+                    LazyColumn(
+                        modifier = Modifier.fillMaxSize(),
+                        verticalArrangement = Arrangement.spacedBy(12.dp)
+                    ) {
+                        groupedExercises.forEach { (muscleGroup, items) ->
+                            val isCollapsed = collapsedStates[muscleGroup] ?: true
+                            stickyHeader {
+                                Surface(
+                                    modifier = Modifier
+                                        .fillMaxWidth()
+                                        .clickable {
+                                            collapsedStates[muscleGroup] = !isCollapsed
+                                        },
+                                    color = MaterialTheme.colorScheme.surface,
+                                    tonalElevation = 4.dp
+                                ) {
+                                    Text(
+                                        text = if (isCollapsed) "▸ $muscleGroup" else "▾ $muscleGroup",
+                                        modifier = Modifier.padding(vertical = 8.dp, horizontal = 16.dp),
+                                        style = MaterialTheme.typography.titleMedium.copy(fontFamily = FontFamily.Serif)
+                                    )
+                                }
+                            }
+                            if (!isCollapsed) {
+                                items(items) { ex ->
+                                    ExerciseCardWithHighlight(
+                                        ex,
+                                        "",
+                                        onEdit = {
+                                            editing = ex
+                                            showSheet = true
+                                        },
+                                        onDelete = { vm.delete(ex.id) }
+                                    )
+                                }
                             }
                         }
                     }
                 }
             }
+        }
+        Button(
+            onClick = { editing = null; showSheet = true },
+            modifier = Modifier
+                .align(Alignment.BottomEnd)
+                .padding(16.dp)
+        ) {
+            Text("➕ Add", fontFamily = FontFamily.Serif)
         }
     }
 
@@ -197,85 +208,5 @@ fun ExerciseManagementScreen(navController: NavController) {
             categories = categories,
             muscleGroups = muscles
         )
-    }
-}
-
-@Composable
-fun ExerciseCardWithHighlight(
-    ex: Exercise,
-    query: String,
-    onEdit: () -> Unit,
-    onDelete: () -> Unit
-) {
-    Card(
-        modifier = Modifier.fillMaxWidth(),
-        shape = RoundedCornerShape(12.dp),
-        elevation = CardDefaults.cardElevation(4.dp),
-        colors = CardDefaults.cardColors(
-            containerColor = MaterialTheme.colorScheme.surfaceVariant
-        )
-    ) {
-        Column(modifier = Modifier.padding(16.dp)) {
-            Text(
-                text = highlightQuery(ex.name, query),
-                style = MaterialTheme.typography.titleMedium,
-                fontFamily = FontFamily.Serif,
-                maxLines = 1,
-                overflow = TextOverflow.Ellipsis
-            )
-            Text(
-                text = "${ex.muscleGroup.display} · ${ex.category.display}",
-                style = MaterialTheme.typography.bodyMedium
-            )
-
-            Spacer(modifier = Modifier.height(8.dp))
-
-            Row(
-                modifier = Modifier.fillMaxWidth(),
-                horizontalArrangement = Arrangement.End
-            ) {
-                TextButton(onClick = onEdit) {
-                    Text("Edit", fontFamily = FontFamily.Serif)
-                }
-                TextButton(onClick = onDelete) {
-                    Text("Delete", fontFamily = FontFamily.Serif)
-                }
-            }
-        }
-    }
-}
-
-fun highlightQuery(text: String, query: String): AnnotatedString {
-    if (query.isBlank()) return AnnotatedString(text)
-    val normalizedText = text.replace("\\s".toRegex(), "").lowercase()
-    val start = normalizedText.indexOf(query)
-    if (start < 0) return AnnotatedString(text)
-
-    var actualStart = 0
-    var currentIndex = 0
-    var normalizedIndex = 0
-    while (normalizedIndex < start && currentIndex < text.length) {
-        if (!text[currentIndex].isWhitespace()) {
-            normalizedIndex++
-        }
-        currentIndex++
-    }
-    actualStart = currentIndex
-
-    var matchCount = 0
-    var endIndex = actualStart
-    while (endIndex < text.length && matchCount < query.length) {
-        if (!text[endIndex].isWhitespace()) {
-            matchCount++
-        }
-        endIndex++
-    }
-
-    return buildAnnotatedString {
-        append(text.substring(0, actualStart))
-        withStyle(SpanStyle(color = Color.Red)) {
-            append(text.substring(actualStart, endIndex))
-        }
-        append(text.substring(endIndex))
     }
 }
